+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Updated `PowEmailConfirmation.Ecto.Schema.confirm_email_changeset/1` so now `:email_confirmation_token` is set to `nil`
 * Fixed bug in `PowEmailConfirmation.Phoenix.ControllerCallbacks.send_confirmation_email/2` where the confirmation e-mail wasn't send to the updated e-mail address
 * Added `PowEmailConfirmation.Ecto.Context.current_email_unconfirmed?/2` and `PowEmailConfirmation.Plug.pending_email_change?/1`
+* Added `:email_validator` configuration option to `Pow.Ecto.Schema.Changeset`
+* Added `Pow.Ecto.Schema.Changeset.validate_email/1`
 
 ## v1.0.12 (2019-08-16)